hw5 bonus

+++ scrapper/src/main/java/edu/java/repository/ChatLinkRepository.java
@@ -7,6 +7,6 @@ import java.util.List;
 public interface ChatLinkRepository {
     List<LinkEntity> findLinksByChatId(long id);
     List<ChatEntity> findChatsByLinkId(int id);
-    void create(long chatId,long linkId);
-    void remove(long chatId,long linkId);
+    void create(long chatId,int linkId);
+    void remove(long chatId,int linkId);
 }
